use mongo driver with atlas mongodb deploy heroku

+++ repository/repository.go
@@ -5,8 +5,8 @@ import "vieclamit/models"
 // Repository interface
 type Repository interface {
 	Insert(recruitment models.Recruitment, collection string) error
-	Delete(collection string) (int, error)
-	FindByUrl(urlJob, collection string) (int, error)
+	Delete(collection string) (int64, error)
+	FindByUrl(urlJob, collection string) (int64, error)
 	FindByLocation(location, collection string) (*models.Recruitments, error)
 	FindBySkill(skill, collection string) (*models.Recruitments, error)
 	FindByCompany(company, collection string) (*models.Recruitments, error)
